Add setFilters function to FileDetails and RequirementFlyout
