+++ plugins/main/public/components/agents/fim/inventory/fileDetail.tsx
@@ -52,6 +52,7 @@ import { RedirectAppLinks } from '../../../../../../../src/plugins/opensearch_da
 import TechniqueRowDetails from '../../../overview/mitre/framework/components/techniques/components/flyout-technique/technique-row-details';
 import { DATA_SOURCE_FILTER_CONTROLLED_CLUSTER_MANAGER } from '../../../../../common/constants';
 import NavigationService from '../../../../react-services/navigation-service';
+import { setFilters } from "../../../common/search-bar/set-filters";
 
 export class FileDetails extends Component {
   props!: {
@@ -586,6 +587,8 @@ export class FileDetails extends Component {
 
           this.discoverFilterManager.addFilters(newFilter);
         }}
+        filters={[]}
+        setFilters={setFilters(this.discoverFilterManager)}
       />
     );
   }
